requirements.txt
gymnasium
pytorch==2.1
matplotlib
numpy
pandas
gymnasium-box2d
conda-forge::box2d-py
